test(scheduler): move reduced budget reclaim test
