+++ pkg/scheduler/actions/reclaim/reclaim_benchmark_test.go
@@ -6,20 +6,14 @@ package reclaim_test
 import (
 	"fmt"
 	"testing"
-	"time"
 
-	"github.com/stretchr/testify/require"
 	. "go.uber.org/mock/gomock"
 	"gopkg.in/h2non/gock.v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	kaiv1 "github.com/kai-scheduler/KAI-scheduler/pkg/apis/kai/v1"
 	kaiv1alpha1 "github.com/kai-scheduler/KAI-scheduler/pkg/apis/kai/v1alpha1"
-	commonconstants "github.com/kai-scheduler/KAI-scheduler/pkg/common/constants"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/actions/reclaim"
-	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/common_info"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/pod_status"
-	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/podgroup_info"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/podgroup_info/subgroup_info"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/topology_info"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/constants"
@@ -51,35 +45,6 @@ const (
 	unschedulableDistributedJobName = "unschedulable-distributed-job"
 	unschedulableDistributedRackKey = "benchmark.kai.scheduler/rack"
 )
-
-func TestReducedBudgetFailedReclaimRecordsScenarioSearchUnresolved(t *testing.T) {
-	defer gock.Off()
-
-	test_utils.InitTestingInfrastructure()
-	controller := NewController(t)
-	defer controller.Finish()
-
-	topology := buildUnschedulableDistributedReclaimBenchmarkTopology(
-		defaultUnschedulableDistributedReclaimBenchmarkParams(10),
-	)
-	ssn := test_utils.BuildSession(topology, controller)
-	ssn.Config.ScenarioSearchBudgets = &kaiv1.ScenarioSearchBudgets{
-		MaxActionSearchDuration: map[string]metav1.Duration{
-			commonconstants.ActionReclaim: {Duration: 250 * time.Millisecond},
-		},
-		MaxJobSearchDuration: &metav1.Duration{Duration: time.Second},
-		MinJobSearchDuration: &metav1.Duration{Duration: 500 * time.Millisecond},
-	}
-
-	reclaim.New().Execute(ssn)
-
-	job := ssn.ClusterInfo.PodGroupInfos[common_info.PodGroupID("unschedulable-distributed-job")]
-	require.NotNil(t, job)
-	require.Empty(t, job.JobFitErrors)
-	require.NotNil(t, job.ScenarioSearchUnresolved)
-	require.Equal(t, podgroup_info.ScenarioSearchResultGeneratorsExhausted, job.ScenarioSearchUnresolved.Reason)
-	require.True(t, job.ScenarioSearchUnresolved.ReducedBudget)
-}
 
 func BenchmarkReclaimUnschedulableDistributedJob_10Node(b *testing.B) {
 	benchmarkReclaimUnschedulableDistributedJob(b, 10)
